fix: migration examples page
