+++ app/examples/page.tsx
@@ -1,0 +1,10 @@
+import { Metadata } from 'next';
+import ExamplesPage from './example-page';
+
+export const metadata: Metadata = {
+  title: 'Examples · Yorkie',
+};
+
+export default async function Page() {
+  return <ExamplesPage />;
+}
