+++ app/examples/profile-stack/profile-stack-page.tsx
@@ -1,5 +1,6 @@
+'use client';
+
 import { NextPage } from 'next';
-import Head from 'next/head';
 import { ExampleLayout } from '@/components';
 import {
   Sidebar,
@@ -18,9 +19,6 @@ const ProfileStackExampleView: NextPage = () => {
     <ExampleLayout breadcrumbTitle={exampleTitle}>
       {() => (
         <>
-          <Head>
-            <title>{`${exampleTitle} · Yorkie Examples`}</title>
-          </Head>
           <Sidebar wide>
             <Sidebar.Tabs defaultTab="code">
               <Sidebar.Top>
